Extract MypageTableViewItem into another file
